HDDS-2408 - Separate OM Storage initialization from OzoneManager - Fixed checkstyle issues, added documentation to new package and classes, and added copyright notice.
