+++ hadoop-ozone/ozone-manager/src/main/java/org/apache/hadoop/ozone/om/OzoneManagerStarter.java
@@ -21,7 +21,7 @@ import org.apache.hadoop.hdds.cli.GenericCli;
 import org.apache.hadoop.hdds.cli.HddsVersionProvider;
 import org.apache.hadoop.hdds.conf.OzoneConfiguration;
 import org.apache.hadoop.hdds.tracing.TracingUtil;
-import org.apache.hadoop.ozone.om.init.OzoneManagerInitializer;
+import org.apache.hadoop.ozone.om.init.OzoneManagerStorageInitializer;
 import org.apache.hadoop.security.authentication.client.AuthenticationException;
 import org.apache.hadoop.util.StringUtils;
 import org.slf4j.Logger;
@@ -124,7 +124,7 @@ public class OzoneManagerStarter extends GenericCli {
 
     public boolean init(OzoneConfiguration conf) throws IOException,
         AuthenticationException {
-      return OzoneManagerInitializer.run(conf);
+      return OzoneManagerStorageInitializer.run(conf);
     }
   }
