+++ hadoop-ozone/ozone-manager/src/main/java/org/apache/hadoop/ozone/om/init/FullSecureInitializer.java
@@ -1,3 +1,20 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with this
+ * work for additional information regarding copyright ownership.  The ASF
+ * licenses this file to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
 package org.apache.hadoop.ozone.om.init;
 
 import org.apache.hadoop.hdds.conf.OzoneConfiguration;
@@ -13,7 +30,7 @@ class FullSecureInitializer extends SecurityInitializer {
 
   @Override
   void initialize(OzoneConfiguration conf) throws IOException {
-    new FullUnsecureInitializer(storage).initialize(conf);
-    new SecurityInitializer(storage).initialize(conf);
+    new FullUnsecureInitializer(getStorage()).initialize(conf);
+    new SecurityInitializer(getStorage()).initialize(conf);
   }
 }
