+++ hadoop-ozone/ozone-manager/src/main/java/org/apache/hadoop/ozone/om/init/SecurityInitializer.java
@@ -1,9 +1,26 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with this
+ * work for additional information regarding copyright ownership.  The ASF
+ * licenses this file to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
 package org.apache.hadoop.ozone.om.init;
 
 import org.apache.hadoop.hdds.HddsUtils;
 import org.apache.hadoop.hdds.conf.OzoneConfiguration;
 import org.apache.hadoop.hdds.protocol.proto.HddsProtos;
-import org.apache.hadoop.hdds.protocol.proto.SCMSecurityProtocolProtos;
+import org.apache.hadoop.hdds.protocol.proto.SCMSecurityProtocolProtos.SCMGetCertResponseProto;
 import org.apache.hadoop.hdds.protocolPB.SCMSecurityProtocolClientSideTranslatorPB;
 import org.apache.hadoop.hdds.security.x509.SecurityConfig;
 import org.apache.hadoop.hdds.security.x509.certificate.client.CertificateClient;
@@ -27,7 +44,7 @@ import static org.apache.hadoop.ozone.OzoneConsts.RPC_PORT;
 import static org.apache.hadoop.ozone.om.OMConfigKeys.OZONE_OM_KERBEROS_KEYTAB_FILE_KEY;
 import static org.apache.hadoop.ozone.om.OMConfigKeys.OZONE_OM_KERBEROS_PRINCIPAL_KEY;
 
-class SecurityInitializer extends OzoneManagerInitializer {
+class SecurityInitializer extends OzoneManagerStorageInitializer {
 
   SecurityInitializer(OMStorage storage){
     super(storage);
@@ -45,8 +62,8 @@ class SecurityInitializer extends OzoneManagerInitializer {
 
   @Override
   void initialize(OzoneConfiguration conf) throws IOException {
-    initializeSecurity(conf, storage);
-    storage.persistCurrentState();
+    initializeSecurity(conf, getStorage());
+    getStorage().persistCurrentState();
   }
 
 
@@ -133,7 +150,7 @@ class SecurityInitializer extends OzoneManagerInitializer {
     SCMSecurityProtocolClientSideTranslatorPB secureScmClient =
         HddsUtils.getScmSecurityClient(config);
 
-    SCMSecurityProtocolProtos.SCMGetCertResponseProto response = secureScmClient.
+    SCMGetCertResponseProto response = secureScmClient.
         getOMCertChain(omDetailsProto, getEncodedString(csr));
     String pemEncodedCert = response.getX509Certificate();
